test: move log service tests into feature module
